Rename committee session service module
